# perf: aminocompat: package to allow optimizations iff amino-compatible